requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
croniter==1.4.1
datasets==2.14.5
dateutils==0.6.12
deepdiff==6.5.0
dill==0.3.7
docker-pycreds==0.4.0
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
h11==0.14.0
huggingface-hub==0.17.1
hydra-core==1.3.2
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lightning==2.0.9
lightning-cloud==0.5.38
lightning-utilities==0.9.0
lit==15.0.7
lmdb==1.4.1
lpips==0.1.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
monai==1.2.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.15
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
ordered-set==4.1.0
packaging==23.1
pandas==2.1.0
pathtools==0.1.2
Pillow==9.3.0
protobuf==4.24.3
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-fid==0.3.0
pytorch-lightning==2.0.9
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.2
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
starsessions==1.3.0
sympy==1.11.1
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==1.1.2
torchvision==0.15.2+cu118
tqdm==4.66.1
traitlets==5.10.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.23.2
wandb==0.15.10
wcwidth==0.2.6
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==2.3.7
xxhash==3.3.0
yarl==1.9.2
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
croniter==1.4.1
datasets==2.14.5
dateutils==0.6.12
deepdiff==6.5.0
dill==0.3.7
docker-pycreds==0.4.0
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
h11==0.14.0
huggingface-hub==0.17.1
hydra-core==1.3.2
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lightning==2.0.9
lightning-cloud==0.5.38
lightning-utilities==0.9.0
lit==15.0.7
lmdb==1.4.1
lpips==0.1.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
monai==1.2.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.15
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
ordered-set==4.1.0
packaging==23.1
pandas==2.1.0
pathtools==0.1.2
Pillow==9.3.0
protobuf==4.24.3
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-fid==0.3.0
pytorch-lightning==2.0.9
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.2
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
starsessions==1.3.0
sympy==1.11.1
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==1.1.2
torchvision==0.15.2+cu118
tqdm==4.66.1
traitlets==5.10.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.23.2
wandb==0.15.10
wcwidth==0.2.6
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==2.3.7
xxhash==3.3.0
yarl==1.9.2